Change operate() calculation method
The previous method of building a function tree and executing it with
operate() had many drawbacks, like the inability to use backspace.
I changed it so the calculations are made parsing the text, dividing
the numbers into a multi-array 'tokens' where each 'layer' represents an
operation.
Them, the code goes through the array making the calculations of each
layer in the correct order of precendence